server side refactor
