+++ docker-compose.yml
@@ -1,0 +1,5 @@
+services:
+  server:
+    image: server
+    ports:
+      - 5500:5500
