--- server/server.js
@@ -1,22 +1,0 @@
-const mongoose = require('mongoose');
-const dotenv = require('dotenv');
-const app = require('./app');
-
-dotenv.config({ path: './config.env' });
-
-const DB = process.env.MONGODB_URI.replace(
-   '<password>',
-   process.env.DATABASE_PASSWORD
-);
-
-mongoose
-   .connect(DB, {
-      useNewUrlParser: true,
-      useUnifiedTopology: true,
-   })
-   .then(console.log('DB connection complete'));
-
-const port = process.env.PORT || 8000;
-app.listen(port, () => {
-   console.log(`Listening on port ${port}...`);
-});
